Gameplay2_v3
The end-game condition is fixed. The previous version stop the game when totalCircle go beyond 9, but it's wrong. This version decide the game end when the player run out of circle, or her total priority larger than or equal 11, or her maximum priority do not larger than opponent exist minimum.